complete auth service

+++ src/users/users.service.ts
@@ -1,42 +1,43 @@
-import { CreateUserDto } from './dtos/create-user.dto';
 import { Injectable, NotFoundException } from '@nestjs/common';
-import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
+import { InjectRepository } from '@nestjs/typeorm';
 import { User } from './user.entity';
 
 @Injectable()
 export class UsersService {
   constructor(@InjectRepository(User) private repo: Repository<User>) {}
 
-  async createUser(data: CreateUserDto) {
-    const user = this.repo.create(data);
+  create(email: string, password: string) {
+    const user = this.repo.create({ email, password });
+
     return this.repo.save(user);
   }
 
-  async findOne(id: number) {
-    const user = await this.repo.findOne({ where: { id: id } });
+  findOne(id: number) {
+    if (!id) {
+      return null;
+    }
+    return this.repo.findOne({ where: { id: id } });
+  }
+
+  find(email: string) {
+    return this.repo.find({ where: { email: email } });
+  }
+
+  async update(id: number, attrs: Partial<User>) {
+    const user = await this.findOne(id);
     if (!user) {
       throw new NotFoundException('user not found');
     }
-    return user;
-  }
-
-  async find(email: string) {
-    const data = await this.repo.find({ where: { email: email } });
-    return data;
-  }
-
-  async updataUser(id: number, data: Partial<User>) {
-    const user = await this.findOne(id);
-
-    Object.assign(user, data);
-
+    Object.assign(user, attrs);
     return this.repo.save(user);
   }
 
-  async removeUser(id: number) {
+  async remove(id: number) {
     const user = await this.findOne(id);
-
-    return await this.repo.remove(user);
+    if (!user) {
+      throw new NotFoundException('user not found');
+    }
+    return this.repo.remove(user);
   }
 }
